+++ swarm-derive/CHANGELOG.md
@@ -4,7 +4,11 @@
   with the new `on_swarm_event` and `on_connection_handler_event`.
   See [PR 3011].
 
+- Add `prelude` configuration option.
+  The derive-macro generates code that needs to refer to various symbols. See [PR 3055].
+
 [PR 3011]: https://github.com/libp2p/rust-libp2p/pull/3011
+[PR 3055]: https://github.com/libp2p/rust-libp2p/pull/3055
 
 # 0.30.1