+++ swarm-derive/CHANGELOG.md
@@ -1,8 +1,13 @@
-# 0.30.2 [unreleased]
+# 0.30.2 - [unreleased]
+
+- Replace `NetworkBehaviour` Derive macro deprecated `inject_*` method implementations
+  with the new `on_swarm_event` and `on_connection_handler_event`.
+  See [PR 3011].
 
 - Add `prelude` configuration option.
   The derive-macro generates code that needs to refer to various symbols. See [PR 3055].
 
+[PR 3011]: https://github.com/libp2p/rust-libp2p/pull/3011
 [PR 3055]: https://github.com/libp2p/rust-libp2p/pull/3055
 
 # 0.30.1